+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [feat]: `magnitude` and `magnitudeSq` properties;
 - [feat]: `unit` property;
 - [feat]: `.setMagnitude()` method;
+- [feat]: `.limit()` method;
 
 ## 0.3.0 (2022-11-25)